+++ src/main/kotlin/io/github/nomisrev/service/DatabasePool.kt
@@ -1,0 +1,6 @@
+package io.github.nomisrev.service
+
+interface DatabasePool {
+  fun isRunning(): Boolean
+  suspend fun version(): String?
+}
