+++ src/test/kotlin/io/github/nomisrev/HealthCheckSpec.kt
@@ -1,5 +1,8 @@
 package io.github.nomisrev
 
+import io.github.nomisrev.config.envConfig
+import io.github.nomisrev.config.module
+import io.github.nomisrev.routes.HealthCheck
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.ktor.client.request.get
